Clean up ClusteringTest
